break up transport tests
